Janitor patch: Fix a bunch of signed-vs-unsigned comparisons
Only for the core so far (or for what gets compiled into the core),
and only the safe ones. A few remain in compound evaluations where
I'd rather not touch anything, since arithmetic on unsigned values
could go negative by design.

Anyways, this makes build output with near-paranoid compiler
warnings almost pleasant to look at.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2737 5f96b256-904b-4d8d-8c98-d829582c6739

+++ include/objects.h
@@ -93,9 +93,9 @@ NAGIOS_BEGIN_DECL
 #define OPT_DISABLED      (1 << 15) /* will denote disabled checks some day */
 
 /* macros useful with both hosts and services */
-#define flag_set(c, flag)    (c |= (flag))
-#define flag_get(c, flag)    (c & (flag))
-#define flag_isset(c, flag)  (flag_get((c), (flag)) == (flag))
+#define flag_set(c, flag)    ((c) |= (flag))
+#define flag_get(c, flag)    (unsigned int)((c) & (flag))
+#define flag_isset(c, flag)  (flag_get((c), (flag)) == (unsigned int)(flag))
 #define flag_unset(c, flag)  (c &= ~(flag))
 #define should_stalk(o) flag_isset(o->stalking_options, 1 << o->current_state)
 #define should_flap_detect(o) flag_isset(o->flap_detection_options, 1 << o->current_state)
